Refactoring insering transacription service
