+++ src/main/java/org/hackathon/openassets/db/repository/mongodb/MappedDocumentRepository.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.hackathon.openassets.db.repository.MappedDocumentRepository;
 import org.hackathon.openassets.model.MappedDocument;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,11 +12,27 @@ import com.google.gson.Gson;
 import com.mongodb.DBObject;
 import com.mongodb.util.JSON;
 
-public class MappedDocumentRepositoryImpl implements MappedDocumentRepository {
+public class MappedDocumentRepository {
 	private final static Logger LOG = LoggerFactory
-			.getLogger(MappedDocumentRepositoryImpl.class);
+			.getLogger(MappedDocumentRepository.class);
 
-	@Override
+	private MappedDocumentDao mappedDocumentsDao;
+
+	public MappedDocumentRepository() {
+		try {
+			DbMongoClient dbMongoClient = new DbMongoClient();
+			mappedDocumentsDao = dbMongoClient.getMappedDocumentsDao();
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
+	}
+	
+	public void insert(String document) {
+		mappedDocumentsDao.insert(document);
+	}
+
+	//?? no longer needed
+	
 	public MappedDocument getById(String documentId) {
 		MappedDocument mappedDocument = null;
 
@@ -30,11 +45,10 @@ public class MappedDocumentRepositoryImpl implements MappedDocumentRepository {
 		}
 		return mappedDocument;
 	}
-	
-	@Override
+
 	public List<MappedDocument> getTrusted() {
 		List<DBObject> objects = mappedDocumentsDao.queryTrustedDocuments();
-		
+
 		List<MappedDocument> documents = new ArrayList<MappedDocument>();
 		for (DBObject object : objects) {
 			String objectString = JSON.serialize(object);
@@ -42,34 +56,9 @@ public class MappedDocumentRepositoryImpl implements MappedDocumentRepository {
 			documents.add((MappedDocument) gson.fromJson(objectString,
 					MappedDocument.class));
 		}
-		
+
 		LOG.info("Returning [count={}] truested documents", documents.size());
 		return documents;
-	}
-
-	@Override
-	public void update(MappedDocument document) {
-		mappedDocumentsDao.update(document);
-
-	}
-
-	private MappedDocumentDao mappedDocumentsDao;
-
-	public MappedDocumentRepositoryImpl() {
-		try {
-			DbMongoClient dbMongoClient = new DbMongoClient();
-			mappedDocumentsDao = dbMongoClient.getMappedDocumentsDao();
-		} catch (IOException e) {
-			e.printStackTrace();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	@Override
-	public void save(MappedDocument document) {
-		mappedDocumentsDao.save(document);
-		
 	}
 
 }
